Split out ModuleLoadResult
